+++ src/lib.rs
@@ -54,192 +54,113 @@
 
 pub mod core;
 pub mod collections;
-pub mod node;
+pub mod graph;
 pub mod dot;
-
-#[macro_export]
-macro_rules! node {
-	( $key:expr ) => {
-        {
-            GraphNode::new($key, None)
-        }
-    };
-    ( $key:expr, $param:expr ) => {
-        {
-            GraphNode::new($key, Some($param))
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! graph {
-
-	// s => [e1, e2]
-	( ($K:ty, $N:ty, $E:ty), $($NODE:expr $( => [ $( $EDGE:expr),*] )? )* )
-	=> {
-        {
-			use std::collections::BTreeMap;
-			let mut edges = Vec::<($K, $K)>::new();
-            let mut map = BTreeMap::<$K, GraphNode<$K, $N, $E>>::new();
-            $(
-				$(
-					$(
-						edges.push(($NODE, $EDGE));
-					)*
-				)?
-				let n = node!($NODE);
-                map.insert(n.id().clone(), n);
-            )*
-			for (s, t) in edges {
-				if map.contains_key(&s) && map.contains_key(&t) {
-					let s = map.get(&s).unwrap();
-					let t = map.get(&t).unwrap();
-					connect!(s => t);
-				}
-			}
-            map
-        }
-    };
-}
-
-#[macro_export]
-macro_rules! graph_with_params {
-// (s, T) => [(e1, T), (e2, T)]
-( ($K:ty, $N:ty) => [$E:ty] $(($NODE:expr, $NPARAM:expr) => $( [$( ( $EDGE:expr, $EPARAM:expr) ),*] )? )* )
-	=> {
-		{
-			use std::collections::BTreeMap;
-			let mut edges = Vec::<($K, $K, $E)>::new();
-			let mut map = BTreeMap::<$K, GraphNode<$K, $N, $E>>::new();
-			$(
-				$(
-					$(
-						edges.push(($NODE, $EDGE, $EPARAM));
-					)*
-				)?
-				let n = node!($NODE, $NPARAM);
-				map.insert(n.id().clone(), n);
-			)*
-			for (s, t, param) in edges {
-				if map.contains_key(&s) && map.contains_key(&t) {
-					let s = map.get(&s).unwrap();
-					let t = map.get(&t).unwrap();
-					connect!(s => t, Some(param));
-				}
-			}
-			map
-		}
-	};
-}
-
-#[macro_export]
-macro_rules! connect {
-	( $s:expr => $t:expr ) => {
-        {
-            Node::connect($s, $t, None)
-        }
-    };
-    ( $s:expr => $t:expr, $params:expr ) => {
-        {
-            Node::connect($s, $t, $params)
-        }
-    };
-}
-
-#[derive(Debug, PartialEq, Eq, Clone, std::hash::Hash)]
-enum Space { Galaxy, Star, Planet, BlackHole, Nebula}
+pub mod macros;
 
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::node::*;
+	use crate::graph::*;
 	use Space::*;
 
-	#[test]
-	fn it_works() {
-		let g1 = graph![
-			(usize, &str, usize),
-			0 => [1, 2]
-			1 => [0, 2]
-			2 => [0, 4]
-			3 => [0, 1, 3]
-			4 => [1, 2, 4, 5]
-			5
-		];
+	#[derive(Debug, PartialEq, Eq, Clone, std::hash::Hash)]
+	enum Space { Galaxy, Star, Planet, BlackHole, Nebula}
 
-		impl std::fmt::Display for Space {
-			fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-				match self {
-					Space::Galaxy => write!(f, "Galaxy"),
-					Space::Star => write!(f, "Star"),
-					Space::Planet => write!(f, "Planet"),
-					Space::BlackHole => write!(f, "Black Hole"),
-					Space::Nebula => write!(f, "Nebula"),
-				}
+	impl std::fmt::Display for Space {
+		fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+			match self {
+				Space::Galaxy => write!(f, "Galaxy"),
+				Space::Star => write!(f, "Star"),
+				Space::Planet => write!(f, "Planet"),
+				Space::BlackHole => write!(f, "Black Hole"),
+				Space::Nebula => write!(f, "Nebula"),
 			}
-		}
-
-		let space = graph_with_params![
-
-			(&str, Space) => [f32]
-
-			("Proxima Centauri", Star) =>
-				[
-					("Milky Way", 2873.124),
-					("Andromeda", 3425.24)
-				]
-
-			("Milky Way", Galaxy) =>
-				[
-					("Proxima Centauri", 547.135),
-					("Andromeda", 78873.145)
-				]
-
-			("Andromeda", Galaxy) =>
-				[
-					("Proxima Centauri", 23.442),
-					("TON-512", 25663.156)
-				]
-
-			("Horse Nebula", Nebula) =>
-				[
-					("Proxima Centauri", 46137.124),
-					("Milky Way", 146.312),
-					("Andromeda", 6143.12)
-				]
-
-			("TON-512", BlackHole) =>
-				[
-					("Milky Way", 134.342),
-					("Andromeda", 24834.1432),
-					("Sagittarius A*", 74313.731)
-				]
-
-			("Sagittarius A*", BlackHole) => []
-
-		];
-
-		println!("");
-		println!("Orginal graph:");
-		println!("");
-
-		space.iter().for_each(|(_,node)| {
-			println!("{}", node);
-		});
-
-		let result = space[&"Proxima Centauri"].breadth_first_search(&space[&"Sagittarius A*"]);
-
-		println!("");
-		println!("BFS results:");
-		println!("");
-
-		for edge in result.unwrap() {
-			println!("{} -> {}", edge.source().id(), edge.target().id());
 		}
 	}
 
 	#[test]
-	fn dot_tests() {
+	fn basic1() {
+
+		let space = graph_with_params![
+
+			(String, Space) => [f32]
+
+			(format!("Proxima Centauri"), Star) =>
+				[
+					(format!("Milky Way"), 2873.124),
+					(format!("Andromeda"), 3425.24)
+				]
+
+			(format!("Milky Way"), Galaxy) =>
+				[
+					(format!("Proxima Centauri"), 547.135),
+					(format!("Andromeda"), 78873.145)
+				]
+
+			(format!("Andromeda"), Galaxy) =>
+				[
+					(format!("Proxima Centauri"), 23.442),
+					(format!("TON-512"), 25663.156)
+				]
+
+			(format!("Horse Nebula"), Nebula) =>
+				[
+					(format!("Proxima Centauri"), 46137.124),
+					(format!("Milky Way"), 146.312),
+					(format!("Andromeda"), 6143.12)
+				]
+
+			(format!("TON-512"), BlackHole) =>
+				[
+					(format!("Milky Way"), 134.342),
+					(format!("Andromeda"), 24834.1432),
+					(format!("Sagittarius A*"), 74313.731)
+				]
+
+			(format!("Sagittarius A*"), BlackHole) => []
+
+		];
+
+		println!("\nOriginal graph:\n");
+		for node in &space { println!("{}", node.1); }
+
+		let source = space.get(&format!("Proxima Centauri")).unwrap();
+		let target = space.get(&format!("Sagittarius A*")).unwrap();
+
+		let result = source.breadth_traversal(|_, t, _| {
+			match t == target {
+				false => Traverse::Include,
+				true => Traverse::Terminate,
+			}
+		});
+
+		// let mut res = Vec::new();
+		// if result.len() == 0 && target != result.last().unwrap().target() {
+		// 	println!("Target not found!");
+		// }
+		// else {
+		// 	let w = result.get(result.len() - 1).unwrap();
+		// 	res.push(w.clone());
+		// 	let mut i = 0;
+		// 	for edge in result.iter().rev() {
+		// 		let source = res[i].source();
+		// 		if edge.target() == source {
+		// 			res.push(edge.clone());
+		// 			i += 1;
+		// 		}
+		// 	}
+		// 	res.reverse();
+
+		// 	println!("\nBFS results:\n");
+
+		// 	for edge in res {
+		// 		println!("{} -> {}", edge.source().id(), edge.target().id());
+		// 	}
+		// }
+	}
+
+	fn basic2() {
 		let g1 = graph![
 			(usize, &str, usize),
 			0 => [1, 2]
@@ -249,6 +170,10 @@ mod tests {
 			4 => [1, 2, 4, 5]
 			5
 		];
+	}
+
+	#[test]
+	fn dot_tests() {
 
 		let space = graph_with_params![
 
@@ -275,5 +200,32 @@ mod tests {
 		];
 
 		dot::to_dot_directed(space);
+	}
+
+	#[test]
+	fn lol ()
+	{
+		use rayon::prelude::*;
+		use std::sync::Arc;
+		use rayon::iter::IntoParallelRefIterator;
+		struct Foo<T: Send + Sync + Sized> {
+			data: T,
+			func: Arc<dyn Fn (T) -> T>,
+		}
+
+		unsafe impl<T: Send + Sync + Sized> Sync for Foo<T> {}
+
+		let foos: Vec<Arc<Foo<i32>>> = Vec::new();
+
+		for _ in 0..10 {
+			foos.push(Arc::new(Foo {
+				data: 1,
+				func: Arc::new(|x| x + 1),
+			}));
+		}
+
+		let res = foos.into_par_iter().map(|foo| {
+			true
+		}).collect::<Vec<_>>();
 	}
 }
